api/requirements.txt
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
sniffio==1.3.1
SQLAlchemy==2.0.27
typing_extensions==4.10.0
urllib3==2.2.1
watchdog==4.0.0
Werkzeug==3.0.1
